I hope ts is final

+++ Semester 2/Lab_3/Lab_3_1/Lab_3_1_module.c
@@ -1,3 +1,12 @@
+/*
+* @file Lab_3_1_module.c
+* @author Дарчук Г.С., 515i
+* @date 24.04.2025
+* @brief Лабораторна робота №3, варіант 22
+*
+* Використання масивів. Завдання 1
+*/
+
 #include <stdio.h>
 #include "Lab_3_1_module.h"
 
@@ -14,7 +23,7 @@ void count_repeated_digits(long n, int counted[])
 
 void write_repeated_digits(int arr[])
 {
-    for (long i = 0; i < 10; i++)
+    for (int i = 0; i < 10; i++)
     {
         if (arr[i] > 1) printf("%d (%d разів)\n", i, arr[i]);
     }
